+++ CHANGELOG.md
@@ -1,3 +1,7 @@
-## 0.0.1
+# Changelog
 
-flutter_tiktoken is a flutter offline package for a fast BPE tokeniser for OpenAI models.
+## 0.0.2
+- Breaking change: Renamed `.initata()` to `.initData()`.
+
+## 0.0.1
+- flutter_tiktoken is a flutter offline package for a fast BPE tokeniser for OpenAI models.